requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
attrs==25.4.0
bitarray==3.8.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.0
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
hexbytes==1.3.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
multidict==6.7.0
narwhals==2.11.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parsimonious==0.10.0
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.0
pyarrow==21.0.0
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyunormalize==17.0.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rlp==4.1.0
rpds-py==0.28.0
six==1.17.0
smmap==5.0.2
streamlit==1.51.0
tenacity==9.1.2
toml==0.10.2
toolz==1.1.0
tornado==6.5.2
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
web3==7.14.0
websockets==15.0.1
yarl==1.22.0
